# contract/openapi-panel-v1.json
{
  "components": {
    "schemas": {
...
      "CommandFailureReason": {
        "description": "Why a command did not happen. UC-16 requires the panel to say which.",
        "enum": [
          "device_offline",
          "no_permission",
          "os_refused",
          "ack_timeout",
          "discarded_stale",
          "unsupported",
          "busy"
        ],
        "title": "CommandFailureReason",
        "type": "string"
      },
      "CommandKind": {
        "description": "Server-to-device commands. No ``send_sms`` — SMS is out of scope.",
        "enum": [
          "dial",
          "config",
          "logout",
          "ping",
          "recheck"
        ],
        "title": "CommandKind",
        "type": "string"
      },
      "CommandListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CommandResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "CommandListResponse",
        "type": "object"
      },
      "CommandResponse": {
        "properties": {
          "acked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acked At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "failure_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CommandFailureReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "installation_id": {
            "format": "uuid",
            "title": "Installation Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/CommandKind"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "acked_at - created_at, stored so UC-16's five-second bar is **measured** rather than assumed. R3 flags that bar as possibly unachievable on doze-restricted OEMs; this column is what turns that into a conversation with evidence.",
            "title": "Latency Ms"
          },
          "result_call_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Call Id"
          },
          "sent_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          },
          "status": {
            "$ref": "#/components/schemas/CommandStatus"
          }
        },
        "required": [
          "id",
          "installation_id",
          "kind",
          "status",
          "failure_reason",
          "created_at",
          "sent_at",
          "acked_at",
          "expires_at",
          "result_call_id",
          "latency_ms"
        ],
        "title": "CommandResponse",
        "type": "object"
      },
      "CommandStatus": {
        "enum": [
          "pending",
          "sent",
          "acknowledged",
          "failed",
          "expired"
        ],
        "title": "CommandStatus",
        "type": "string"
      },
...
      "CreateCommandRequest": {
        "description": "``POST /api/v1/devices/{installation_id}/commands``.",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/CommandKind"
          },
          "number": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "For kind='dial'. Named, not a free-form payload map (§8).",
            "title": "Number"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "kind"
        ],
        "title": "CreateCommandRequest",
        "type": "object"
      },
      "CreateDirectoryEntryRequest": {
        "description": "``POST /api/v1/line-directory`` — an admin extra for UC-25.",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/DirectoryRuleKind"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "pattern": {
            "description": "Digits to match. UC-25's '*700' is the suffix rule '700'.",
            "maxLength": 32,
            "minLength": 2,
            "title": "Pattern",
            "type": "string"
          }
        },
        "required": [
          "pattern",
          "kind"
        ],
        "title": "CreateDirectoryEntryRequest",
        "type": "object"
      },
...
      "DirectoryEntryListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DirectoryEntryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "DirectoryEntryListResponse",
        "type": "object"
      },
      "DirectoryEntryResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "kind": {
            "$ref": "#/components/schemas/DirectoryRuleKind"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "pattern": {
            "title": "Pattern",
            "type": "string"
          }
        },
        "required": [
          "id",
          "pattern",
          "kind",
          "label",
          "is_active",
          "created_at"
        ],
        "title": "DirectoryEntryResponse",
        "type": "object"
      },
      "DirectoryRuleKind": {
        "description": "UC-25's ``*700`` is a ``suffix`` rule.",
        "enum": [
          "exact",
          "prefix",
          "suffix"
        ],
        "title": "DirectoryRuleKind",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReclassifyResponse": {
        "description": "A directory change is only half done until the calls agree with it.",
        "properties": {
          "calls_reclassified": {
            "description": "Calls whose call_type changed as a result of this edit.",
            "title": "Calls Reclassified",
            "type": "integer"
          },
          "entry": {
            "$ref": "#/components/schemas/DirectoryEntryResponse"
          }
        },
        "required": [
          "entry",
          "calls_reclassified"
        ],
        "title": "ReclassifyResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Generated from the Pydantic schemas. Prefix: /api/v1",
    "title": "BonviCall panel API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/commands/{command_id}": {
      "get": {
        "description": "Status, latency and the failure reason — UC-16 requires the reason.",
        "operationId": "get_command_api_v1_commands__command_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "command_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Command Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Command",
        "tags": [
          "Commands"
        ]
      }
    },
...
    "/api/v1/devices/{installation_id}/commands": {
      "get": {
        "description": "History with ``latency_ms``, which is how UC-16's bar is measured.",
        "operationId": "list_commands_api_v1_devices__installation_id__commands_get",
        "parameters": [
          {
            "in": "path",
            "name": "installation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Installation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Commands",
        "tags": [
          "Commands"
        ]
      },
      "post": {
        "description": "202: accepted for delivery, not yet done.\n\nIssuing a command reaches into an employee's personally owned phone, so it\nis audited under ``command_issued`` — the only action in the product that\nacts on hardware the company does not own.",
        "operationId": "issue_command_api_v1_devices__installation_id__commands_post",
        "parameters": [
          {
            "in": "path",
            "name": "installation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Installation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue Command",
        "tags": [
          "Commands"
        ]
      }
    },
...
    "/api/v1/line-directory": {
      "get": {
        "operationId": "list_entries_api_v1_line_directory_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryEntryListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Entries",
        "tags": [
          "Line directory"
        ]
      },
      "post": {
        "description": "Add a rule and re-run the classification in the same transaction.",
        "operationId": "create_entry_api_v1_line_directory_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDirectoryEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReclassifyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Entry",
        "tags": [
          "Line directory"
        ]
      }
    },
    "/api/v1/line-directory/{entry_id}": {
      "delete": {
        "description": "Deactivate a rule and re-run the classification.",
        "operationId": "delete_entry_api_v1_line_directory__entry_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReclassifyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Entry",
        "tags": [
          "Line directory"
        ]
      }
    },
...
  }
}